@@ -5,7 +5,7 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
@@ -14,6 +14,7 @@ dependencies:
   permission_handler: ^5.1.0+2
   vector_math: ^2.0.8
   json_annotation: ^3.0.1
+  google_sign_in: ^5.0.1
 
 dev_dependencies:
   flutter_test:
